pyproject.toml
[tool.poetry]
name = "pystockfilter"
version = "1.0.9"
description = "Intelligent Stock Screening for Informed Investment Decisions"
authors = ["SlashGordon <slash.gordon.dev@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dateutil = "^2.8.2"
pandas-ta = "^0.3.14b0"
pystockdb = "^1.2.0"
backtesting = "^0.3.3"

[tool.poetry.group.dev.dependencies]
wheel = "^0.41.3"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
autopep8 = "^2.0.4"
freezegun = "^1.2.2"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
